v0.7.8 release (#107)
* bugfix: numpy v1.26 introduces an error when installing from Github.

As far as I can tell, this comes down to something to do with versioneer in seaborn or pandas not playing nice with dev installs. The best move would probably be to upgrade to numpy>2. But to avoid breaking backwards compatibility, I'm trying out with just excluding numpy v1.26 (i.e. still allowing v1.24 and v1.25 for those that need v1 for whatever reason).

Signed-off-by: Andrew Mitchell <a.j.mitchell@ucl.ac.uk>

* bump version v0.7.6 -> v0.7.7

* bump version v0.7.7 -> v0.7.8dev0

* update to use acoustic-toolbox (#106)

Signed-off-by: Andrew Mitchell <a.j.mitchell@ucl.ac.uk>

* bump version v0.7.8dev0 -> v0.7.8rc0

* bump version v0.7.8rc0 -> v0.7.8rc1

---------

Signed-off-by: Andrew Mitchell <a.j.mitchell@ucl.ac.uk>